menus.rs and font preference
This commit mainly introduces the font preference
menu item. Menu-related code is moved to menus.rs.
IDs are changed to styles such as "f_toc".

For simplicity, menus are always for app, not window.

Font preference is also stored in file prefs.json.
However, this value often does not have an effect
in front-end because of the specific rules per class.